requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
drf-spectacular==0.24.0
drf-spectacular-sidecar==2022.9.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.16.0
Markdown==3.4.1
pkgutil-resolve-name==1.3.10
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.8.1
